+++ pytsite/file/_model.py
@@ -1,11 +1,11 @@
 """File Model.
 """
+from os import path as _path, unlink as _unlink
+from pytsite import odm as _odm, reg as _reg, router as _router
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
-
-from os import path as _path, unlink as _unlink
-from pytsite import odm as _odm, reg as _reg, router as _router
 
 
 class File(_odm.Model):